requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pydantic==1.10.15
pyyaml==6.0.2
loguru==0.7.2
sqlalchemy==2.0.23
numpy>=1.24,<2.0
opencv-python==4.10.0.84
torch>=2.1.0
torchvision>=0.16.0
facenet-pytorch==2.6.0
faiss-cpu==1.7.4

